+++ dao/src/main/java/by/it/academy/dao/IUserDao.java
@@ -1,0 +1,10 @@
+package by.it.academy.dao;
+
+import by.it.academy.model.User;
+
+/**
+ * interface for users
+ */
+public interface IUserDao {
+    User getUserByEmail(String email);
+}
